change to nuxt3
