+++ nuxt.config.ts
@@ -1,0 +1,10 @@
+// https://nuxt.com/docs/api/configuration/nuxt-config
+export default defineNuxtConfig({
+    css: [
+        "vuetify/lib/styles/main.sass",
+        "@mdi/font/css/materialdesignicons.min.css"
+      ],
+    build: {
+        transpile: ["vuetify"],
+    },
+})
