+++ package.json
@@ -1,56 +1,18 @@
 {
-  "name": "vue3-frontend",
-  "version": "0.0.0",
   "private": true,
   "scripts": {
-    "dev": "vite",
-    "build": "run-p type-check build-only",
-    "preview": "vite preview",
-    "test:unit": "vitest",
-    "build-only": "vite build",
-    "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
-    "format": "prettier --write src/",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
-  },
-  "dependencies": {
-    "axios": "^1.3.4",
-    "pinia": "^2.0.32",
-    "vue": "^3.2.47",
-    "vue-axios": "^3.5.2",
-    "vue-router": "^4.1.6",
-    "vuetify": "^3.1.7"
+    "build": "nuxt build",
+    "dev": "nuxt dev",
+    "generate": "nuxt generate",
+    "preview": "nuxt preview",
+    "postinstall": "nuxt prepare"
   },
   "devDependencies": {
+    "nuxt": "^3.2.3"
+  },
+  "dependencies": {
     "@mdi/font": "^7.1.96",
-    "@rushstack/eslint-patch": "^1.2.0",
-    "@storybook/addon-essentials": "^7.0.0-beta.60",
-    "@storybook/addon-interactions": "^7.0.0-beta.60",
-    "@storybook/addon-links": "^7.0.0-beta.60",
-    "@storybook/blocks": "^7.0.0-alpha.8",
-    "@storybook/testing-library": "^0.0.14-next.1",
-    "@storybook/vue3": "^7.0.0-beta.60",
-    "@storybook/vue3-vite": "^7.0.0-beta.60",
-    "@types/jsdom": "^21.1.0",
-    "@types/node": "^18.14.2",
-    "@vitejs/plugin-vue": "^4.0.0",
-    "@vitejs/plugin-vue-jsx": "^3.0.0",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.2",
-    "@vue/test-utils": "^2.3.0",
-    "@vue/tsconfig": "^0.1.3",
-    "eslint": "^8.34.0",
-    "eslint-plugin-vue": "^9.9.0",
-    "jsdom": "^21.1.0",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "storybook": "^7.0.0-beta.60",
-    "typescript": "~4.8.4",
-    "vite": "^4.1.4",
-    "vitest": "^0.29.1",
-    "vue-tsc": "^1.2.0"
+    "sass": "^1.59.2",
+    "vuetify": "^3.1.8"
   }
 }
