+++ plugins/vuetify.ts
@@ -1,0 +1,12 @@
+import * as components from "vuetify/components";
+import * as directives from "vuetify/directives";
+import { createVuetify } from "vuetify";
+
+export default defineNuxtPlugin((nuxt) => {
+  const vuetify = createVuetify({
+    ssr: true,
+    components,
+    directives,
+  });
+  nuxt.vueApp.use(vuetify);
+});
