--- src/stories/Header.stories.ts
@@ -1,42 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/vue3";
-
-import MyHeader from "./Header.vue";
-
-const meta: Meta<typeof MyHeader> = {
-  /* 👇 The title prop is optional.
-   * See https://storybook.js.org/docs/7.0/vue/configure/overview#configure-story-loading
-   * to learn how to generate automatic titles
-   */
-  title: "Example/Header",
-  component: MyHeader,
-  render: (args: any) => ({
-    components: { MyHeader },
-    setup() {
-      return { args };
-    },
-    template: '<my-header :user="args.user" />',
-  }),
-  parameters: {
-    // More on how to position stories at: https://storybook.js.org/docs/7.0/react/configure/story-layout
-    layout: "fullscreen",
-  },
-  // This component will have an automatically generated docsPage entry: https://storybook.js.org/docs/7.0/vue/writing-docs/docs-page
-  tags: ["autodocs"],
-};
-
-export default meta;
-type Story = StoryObj<typeof MyHeader>;
-
-export const LoggedIn: Story = {
-  args: {
-    user: {
-      name: "Jane Doe",
-    },
-  },
-};
-
-export const LoggedOut: Story = {
-  args: {
-    user: null,
-  },
-};
